+++ web-client/src/components/forms/Dropdown/Dropdown.rendering.test.tsx
@@ -1,0 +1,9 @@
+import { render, screen } from "@testing-library/react"
+import { Dropdown } from "./Dropdown.component"
+import { nanoid } from "nanoid"
+
+describe("Dropdown.component: rendering", () => {
+  describe("when the component is rendered by default", () => {
+    it("renders with empty string", () => {})
+  })
+})