+++ web-client/src/components/forms/index.ts
@@ -1,1 +1,4 @@
 export { Button } from "./Button/Button.component"
+export { Checkbox } from "./Checkbox/Checkbox.component"
+export { TextField } from "./TextField/TextField.component"
+export { Dropdown } from "./Dropdown/Dropdown.component"